refactor(data): migrate data to separate folder
