+++ src/pages/events.astro
@@ -1,7 +1,7 @@
 ---
 import BaseLayout from "@/layouts/BaseLayout.astro";
 import EventContainer from "@/components/Event/EventContainer";
-import { EVENTS } from "@/components/Event/TestData";
+import { EVENTS } from "@/data/events";
 ---
 
 <BaseLayout>
