+++ src/pages/faqs/en.astro
@@ -2,7 +2,7 @@
 import BaseLayout from "@/layouts/BaseLayout.astro";
 import TextBox from "@/components/FAQs/TextBox.astro";
 import type { Props as TextBoxProps } from "@/components/FAQs/TextBox.astro";
-import { FAQsEng } from "@/components/FAQs/FAQsData";
+import { FAQS_ENG } from "@/data/FAQs";
 ---
 
 <BaseLayout withJigsaw={false}>
@@ -34,7 +34,7 @@ import { FAQsEng } from "@/components/FAQs/FAQsData";
     </section>
     <div class="w-full max-w-5xl">
       {
-        FAQsEng.map(({ question, answer }: TextBoxProps) => {
+        FAQS_ENG.map(({ question, answer }: TextBoxProps) => {
           return <TextBox question={question} answer={answer} />;
         })
       }
